Renamed send_strategies -> sendstrategies, java guidelines
